refactor : constants 패키지 생성 및 ServiceStatus 이동
